My Resident Evil 5:  Rocket Calculator 1.0.0
A simple application to calculate level times in the Resident Evil 5 game
to determine if you have reached the five hour requirement to unlock the
infinite rocket launchers.  If you’ve exceeded the five hour requirement
the program will tell you how much time you need to shave off.  I’m
working on a way to create a save file (as seen in the source) but it isn’t
working yet.